+++ CHANGELOG.md
@@ -10,10 +10,16 @@ about writing changes to this log.
 
 ## [Unreleased]
 
+- [TVIST1-436](https://jira.itkdev.dk/browse/TVIST1-436)
+  Made `CaseEvent.createdBy` not nullable.
+  Marked `Document.originalFileName` as nullable
 - [TVIST1-743](https://jira.itkdev.dk/browse/TVIST1-743)
   Added case events.
 - [TVIST1-747](https://jira.itkdev.dk/browse/TVIST1-747)
   Added suffix to document names when uploading multiple in one go.
+- [TVIST1-788](https://jira.itkdev.dk/browse/TVIST1-788)
+  - Updated `CaseEvent` with `senders` and `recipients` property.
+  - Removed `CaseEventPartyRelation`.
 
 ## [1.1.2] 2023-02-17